requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-ldap==3.4.4
SQLAlchemy==2.0.31
typing_extensions==4.9.0
Werkzeug==3.0.3
email-validator==2.1.0.post1
mysqlclient==2.2.4